+++ ChangeLog.md
@@ -46,6 +46,7 @@
 - Add build_mscos.sh. add App bundle
 - Add run the program directly without installation for windows install package
 - [Add store password to system redential](https://github.com/KangLin/RabbitRemoteControl/issues/78)
+- Plugin: fix disable desktop shortcuts in linux. See: https://github.com/KangLin/RabbitRemoteControl/discussions/91
 
 ### v0.0.36
 - Client: